Split architecture checker support modules
Move policy and CLI wiring out of the core checker, and isolate barrel
re-export parsing in its own module. Preserve compatibility imports while
resolving nested star-only re-export origins correctly.
