+++ episodic/architecture/__main__.py
@@ -1,6 +1,6 @@
 """Command-line entrypoint for architecture enforcement."""
 
-from .checker import main
+from .cli import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
