+++ tests/fixtures/architecture/api_imports_nested_star_reexported_outbound_adapter/__init__.py
@@ -1,3 +1,3 @@
 """Fixture package that star-re-exports an outbound adapter."""
 
-from .storage import *  # noqa: F403
+from .barrel import *  # noqa: F403
