+++ tests/fixtures/architecture/api_imports_nested_star_reexported_outbound_adapter/barrel.py
@@ -1,0 +1,3 @@
+"""Star-only barrel that forwards public storage exports."""
+
+from .storage import *  # noqa: F403
