+++ tests/fixtures/architecture/api_imports_nested_star_reexported_outbound_adapter/storage.py
@@ -1,7 +1,5 @@
 """Outbound storage adapter re-exported through a star import."""
 
-__all__ = ["StorageAdapter"]
-
 
 class StorageAdapter:
     """Concrete outbound adapter used to test nested package star re-exports."""
